feat: Add models for various scenarios including nested objects, custom mapping and simple data structures
