+++ CrossTypeExpressionConverter.Tests/Helpers/Models/DestDifferentNames.cs
@@ -1,0 +1,10 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class DestDifferentNames
+{
+    public int EntityId { get; set; }
+    public string? FullName { get; set; }
+    public bool Enabled { get; set; }
+    public double NumericValue { get; set; }
+    public DateTime Timestamp { get; set; }
+}
