+++ CrossTypeExpressionConverter.Tests/Helpers/Models/DestSimple.cs
@@ -1,0 +1,10 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class DestSimple
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public bool IsActive { get; set; }
+    public double Value { get; set; }
+    public DateTime CreatedDate { get; set; }
+}
