+++ CrossTypeExpressionConverter.Tests/Helpers/Models/NestedDestProp.cs
@@ -1,0 +1,7 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class NestedDestProp
+{
+    public int NestedId { get; set; }
+    public string? NestedName { get; set; }
+}
