+++ CrossTypeExpressionConverter.Tests/Helpers/Models/NestedSourceProp.cs
@@ -1,0 +1,7 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class NestedSourceProp
+{
+    public int NestedId { get; set; }
+    public string? NestedName { get; set; }
+}
