+++ CrossTypeExpressionConverter.Tests/Helpers/Models/SourceForCustomMap.cs
@@ -1,0 +1,8 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class SourceForCustomMap
+{
+    public int Id { get; set; }
+    public string? Data { get; set; }
+    public int NumericValue { get; set; }
+}
