+++ CrossTypeExpressionConverter.Tests/Helpers/Models/SourceWithNested.cs
@@ -1,0 +1,8 @@
+namespace CrossTypeExpressionConverter.Tests.Helpers.Models;
+
+public class SourceWithNested
+{
+    public int ParentId { get; set; }
+    public NestedSourceProp? Child { get; set; }
+    public NestedSourceProp? ChildToMap { get; set; }
+}
